requirements.txt
jieba==0.42.1
matplotlib==3.6.3
numpy==1.23.4
pandas==1.5.1
prophet==1.1.2
python_dateutil==2.8.2
scikit_learn==1.2.1
sentence_transformers==2.2.2
tensorboardX==2.6
torch==1.11.0+cu102
tqdm==4.64.1
transformers==4.26.0
